requirements.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython==0.29.16
decorator==4.4.1
defusedxml==0.6.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
grpcio==1.27.2
h5py==2.10.0
idna==2.9
ignite==1.1.0
imageio==2.8.0
importlib-metadata==1.3.0
ipdb==0.13.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==5.0.3
networkx==2.4
notebook==6.0.2
numpy==1.17.0
oauthlib==3.1.0
opencv-python==4.1.2.30
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
progressbar==2.5
progressbar2==3.50.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-utils==2.4.0
pytorch-ignite==0.3.0
pytrec-eval==0.4
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
SimpleITK==1.2.4
six==1.13.0
sklearn==0.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.6
zipp==0.6.0
